backend/requirements.txt
fastapi
uvicorn
supabase
mistralai
pypdf
python-multipart
python-dotenv
